Condense multiple paths sent to prometheus into single routes for rails apps
One controller and action could manage thousands of unique paths which in turn would create
thousands of unique timeseries. This places an unnecessarily high load on prometheus.

This commit replaces absolute paths with references to the controller and action that handled it.
If no route is found then we will continue to send the original path.

This is only for rails apps. Non rails apps will continue to send the original path.

Example
`blogs/8`, `blogs/9` => `blogs#view`

https://trello.com/c/lUibintk/369-cut-down-the-number-of-metrics-exposed-for-dynamic-paths